+++ server/ai_mediaserver.go
@@ -750,8 +750,6 @@ func (ls *LivepeerServer) CreateWhip(server *media.WHIPServer) http.Handler {
 			whepURL = "http://localhost:8889/" // default mediamtx output
 		}
 		whepURL = whepURL + streamName + "-out/whep"
-		streamID := ""
-		pipelineID := ""
 
 		go func() {
 			internalOutputHost := os.Getenv("LIVE_AI_PLAYBACK_HOST") // TODO proper cli arg
@@ -760,6 +758,8 @@ func (ls *LivepeerServer) CreateWhip(server *media.WHIPServer) http.Handler {
 			}
 			mediamtxOutputURL := internalOutputHost + streamName + "-out"
 			outputURL := ""
+			streamID := ""
+			pipelineID := ""
 			pipeline := ""
 			pipelineParams := make(map[string]interface{})
 			sourceTypeStr := "livepeer-whip"
@@ -898,17 +898,45 @@ func (ls *LivepeerServer) CreateWhip(server *media.WHIPServer) http.Handler {
 			if err != nil {
 				stopPipeline(err)
 			}
+
+			statsContext, statsCancel := context.WithCancel(ctx)
+			defer statsCancel()
+			go runStats(statsContext, whipConn, streamID, pipelineID, requestID)
+
 			whipConn.AwaitClose()
 			ssr.Close()
 			cleanupControl(ctx, params)
 			clog.Info(ctx, "Live cleaned up")
 		}()
 
-		conn := server.CreateWHIP(ctx, ssr, whepURL, w, r, func(stats *media.MediaStats) {
-			if streamID == "" || pipelineID == "" {
-				clog.V(common.DEBUG).Info(ctx, "ignoring whip stats, streamID or pipelineID not set yet")
+		conn := server.CreateWHIP(ctx, ssr, whepURL, w, r)
+		whipConn.SetWHIPConnection(conn) // might be nil if theres an error and thats okay
+	})
+}
+func runStats(ctx context.Context, whipConn *media.WHIPConnection, streamID string, pipelineID string, requestID string) {
+	// Periodically check whip stats and write logs and metrics
+	ticker := time.NewTicker(5 * time.Second)
+	defer ticker.Stop()
+	for {
+		select {
+		case <-ctx.Done():
+			return
+		case <-ticker.C:
+			stats, err := whipConn.Stats()
+			if err != nil {
 				return
 			}
+			if monitor.Enabled {
+				monitor.AIWhipTransportBytesReceived(int64(stats.PeerConnStats.BytesReceived))
+				monitor.AIWhipTransportBytesSent(int64(stats.PeerConnStats.BytesSent))
+				monitor.AIWhipTransportPacketsReceived(int64(stats.PeerConnStats.PacketsReceived))
+				monitor.AIWhipTransportPacketsSent(int64(stats.PeerConnStats.PacketsSent))
+			}
+			clog.Info(ctx, "whip TransportStats", "ID", stats.PeerConnStats.ID, "bytes_received", stats.PeerConnStats.BytesReceived, "bytes_sent", stats.PeerConnStats.BytesSent, "packets_received", stats.PeerConnStats.PacketsReceived, "packets_sent", stats.PeerConnStats.PacketsSent)
+			for _, s := range stats.TrackStats {
+				clog.Info(ctx, "whip InboundRTPStreamStats", "kind", s.Kind, "jitter", s.Jitter, "packets_lost", s.PacketsLost, "packets_received", s.PacketsReceived, "rtt", s.RTT)
+			}
+
 			monitor.SendQueueEventAsync("stream_ingest_metrics", map[string]interface{}{
 				"timestamp":   time.Now().UnixMilli(),
 				"stream_id":   streamID,
@@ -916,11 +944,9 @@ func (ls *LivepeerServer) CreateWhip(server *media.WHIPServer) http.Handler {
 				"request_id":  requestID,
 				"stats":       stats,
 			})
-		})
-		whipConn.SetWHIPConnection(conn) // might be nil if theres an error and thats okay
-	})
+		}
+	}
 }
-
 func (ls *LivepeerServer) WithCode(code int) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		corsHeaders(w, r.Method)
